server/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.35.36
botocore==1.35.36
cachecontrol==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
faker==37.4.2
flask==3.1.1
flask-cors==6.0.1
flask-jwt-extended==4.7.1
flask-mail==0.10.0
flask-migrate==4.1.0
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
Pillow==10.4.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
markupsafe==3.0.2
msgpack==1.1.1
multidict==6.6.3
packaging==25.0
pg8000==1.31.4
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyjwt==2.10.1
pyparsing==3.2.3
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
requests==2.32.4
rsa==4.9.1
scramp==1.4.6
setuptools==70.3.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.29
sqlalchemy-serializer==1.4.22
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
werkzeug==3.1.3
yarl==1.20.1
